xsigo: Fix slab-out-of-bounds in vhba_create
allocate the sizeof(unsigned long) instead an int.
memory allocated was insufficient for private data,
private data allocated by scsi_host_alloc since
we specify a private data size of sizeof(int),
but later it don.t take that into account, and directly
set the value like: *(host->hostdata) = (int)vhba->idr;
The declaration of hostdata is: unsigned long hostdata[0]
/* Used for storage of host specific stuff */
__attribute__ ((aligned (sizeof(unsigned long))));
So *host->hostdata is de-referencing an unsigned long instead of an int

[  143.804025] Hardware name: Oracle Corporation ORACLE SERVER X5-2/ASM,MOTHERBOARD,1U, BIOS 30370100 07/12/2022
[  143.804038] Workqueue: xsvhba_wq vhba_xsmp_create [xsvhba]
[  143.804039] Call Trace:
[  143.804046]  dump_stack+0x95/0xca
[  143.804058]  ? vhba_create+0x57b/0xdcc [xsvhba]
[  143.804063]  print_address_description.constprop.7+0x6b/0x3ec
[  143.804075]  ? vhba_create+0x57b/0xdcc [xsvhba]
[  143.804086]  ? vhba_create+0x57b/0xdcc [xsvhba]
[  143.804089]  __kasan_report.cold.10+0x37/0x77
[  143.804100]  ? vhba_create+0x57b/0xdcc [xsvhba]
[  143.804103]  kasan_report+0x14/0x1f
[  143.804106]  __asan_store8+0x78/0xa9
[  143.804117]  vhba_create+0x57b/0xdcc [xsvhba]
[  143.804131]  vhba_xsmp_create+0x5f/0x8b [xsvhba]
[  143.804136]  process_one_work+0x413/0x81d
[  143.804139]  ? __kasan_check_write+0x14/0x1a
[  143.804142]  worker_thread+0x57/0x5ac
[  143.804146]  kthread+0x1c8/0x1e5
[  143.804149]  ? process_one_work.cold.51+0x5b/0x58
[  143.804151]  ? __kthread_cancel_work+0xa0/0x98
[  143.804162]  ret_from_fork+0x2b/0x36

[  143.805825] Allocated by task 9870:
[  143.809718]  save_stack+0x21/0x8b
[  143.809720]  __kasan_kmalloc.constprop.12+0xc8/0xcd
[  143.809722]  kasan_kmalloc+0x9/0xf
[  143.809724]  __kmalloc+0x15d/0x37a
[  143.809728]  scsi_host_alloc+0x55/0x733
[  143.809739]  vhba_create+0x1d5/0xdcc [xsvhba]
[  143.809750]  vhba_xsmp_create+0x5f/0x8b [xsvhba]
[  143.809752]  process_one_work+0x413/0x81d
[  143.809754]  worker_thread+0x57/0x5ac
[  143.809756]  kthread+0x1c8/0x1e5
[  143.809757]  ret_from_fork+0x2b/0x36

[  143.811419] Freed by task 9005:
[  143.814923]  save_stack+0x21/0x8b
[  143.814925]  __kasan_slab_free+0x141/0x1c9
[  143.814927]  kasan_slab_free+0xe/0x14
[  143.814929]  kfree+0xb1/0x4e9
[  143.814932]  kvfree+0x46/0x7c
[  143.814936]  seq_release+0x30/0x48
[  143.814939]  kernfs_fop_release+0x77/0x13c
[  143.814942]  __fput+0x15d/0x398
[  143.814944]  ____fput+0xe/0x14
[  143.814946]  task_work_run+0xa2/0xdb
[  143.814950]  exit_to_usermode_loop+0x1e4/0x1e9
[  143.814952]  do_syscall_64+0x21a/0x23c
[  143.814956]  entry_SYSCALL_64_after_hwframe+0x17b/0x0

[  143.816617] The buggy address belongs to the object at ffff889fb4156000
 which belongs to the cache kmalloc-4k of size 4096
[  143.823482] eth4B Install Success: vnet_mode:1 type:0 eoib[no] HPort:1
[  143.830596] The buggy address is located 2360 bytes inside of
 4096-byte region [ffff889fb4156000, ffff889fb4157000)
[  143.830597] The buggy address belongs to the page:
[  143.830600] page:ffffea007ed05400 refcount:1 mapcount:0 mapping:ffff888107c15ac0 index:0x0 compound_mapcount: 0
[  143.830604] flags: 0x57ffffc0010200(slab|head)
[  143.843906] VNIC:eth4B MTU[9000:4:4] MGID:ff12:401c:0000:fde9:0000:0000:ffff:ffff pkey:65535
[  143.849253] raw: 0057ffffc0010200 ffffea007f16c000 0000000200000002 ffff888107c15ac0
[  143.849255] raw: 0000000000000000 0000000080040004 00000001ffffffff 0000000000000000
[  143.849256] page dumped because: kasan: bad access detected

[  143.849270] Memory state around the buggy address:
[  143.849272]  ffff889fb4156800: 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[  143.864345]  ffff889fb4156880: 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
[  143.872408] >ffff889fb4156900: 00 00 00 00 00 00 00 04 fc fc fc fc fc fc fc fc
[  143.880469]                                         ^
[  143.886107]  ffff889fb4156980: fc fc fc fc fc fc fc fc fc fc fc fc fc fc fc fc
[  143.894179]  ffff889fb4156a00: fc fc fc fc fc fc fc fc fc fc fc fc fc fc fc fc
--------

Orabug: 35180168

+++ drivers/infiniband/ulp/xsigo/xsvhba/vhba_create.c
@@ -120,7 +120,7 @@ int vhba_create(xsmp_cookie_t xsmp_hndl, struct vhba_xsmp_msg *msg)
 	if (vhba_max_transfer_size != VHBA_DEFAULT_TRANSFER_SIZE)
 		xg_vhba_driver_template.max_sectors = vhba_max_transfer_size;
 
-	host = scsi_host_alloc(&xg_vhba_driver_template, sizeof(int));
+	host = scsi_host_alloc(&xg_vhba_driver_template, sizeof(unsigned long));
 
 	if (host == NULL) {
 		eprintk(NULL, "scsi host alloc failed\n");
